feat(auth): implement user authentication flow with signup, login, logout, password recovery, and email verification
- Added signup functionality with email verification
- Implemented login with session management and welcome email
- Created logout functionality to clear user session
- Developed password recovery process with token generation and email sending
- Added email verification process with token validation
- Implemented resend verification email functionality
- Created password reset functionality with token validation and email confirmation

feat(user): implement user profile management

- Added functionality to get current user profile
- Implemented user profile update functionality
- Created user account deletion functionality
