+++ apps/client/next.config.mjs
@@ -10,15 +10,24 @@ const nextConfig = {
   //   unoptimized: true,
   // },
   images: {
-   remotePatterns: [
-     {
-       protocol: 'https',
-       hostname: 'cdn.sendexa.co',
-       port: '',
-       pathname: '/**',
-     },
-   ],
+    remotePatterns: [
+      {
+        protocol: "https",
+        hostname: "cdn.sendexa.co",
+        port: "",
+        pathname: "/**",
+      },
+    ],
   },
-}
 
-export default nextConfig
+  async rewrites() {
+    return [
+      {
+        source: "/api/:path*",
+        destination: "http://localhost:4000/api/:path*",
+      },
+    ];
+  },
+};
+
+export default nextConfig;
