+++ apps/client/package.json
@@ -9,72 +9,72 @@
     "start": "next start"
   },
   "dependencies": {
-    "@hookform/resolvers": "^3.10.0",
-    "@radix-ui/react-accordion": "latest",
-    "@radix-ui/react-alert-dialog": "latest",
-    "@radix-ui/react-aspect-ratio": "latest",
-    "@radix-ui/react-avatar": "latest",
-    "@radix-ui/react-checkbox": "latest",
-    "@radix-ui/react-collapsible": "latest",
-    "@radix-ui/react-context-menu": "latest",
-    "@radix-ui/react-dialog": "latest",
-    "@radix-ui/react-dropdown-menu": "latest",
-    "@radix-ui/react-hover-card": "latest",
-    "@radix-ui/react-label": "latest",
-    "@radix-ui/react-menubar": "latest",
-    "@radix-ui/react-navigation-menu": "latest",
-    "@radix-ui/react-popover": "latest",
-    "@radix-ui/react-progress": "latest",
-    "@radix-ui/react-radio-group": "latest",
-    "@radix-ui/react-scroll-area": "latest",
-    "@radix-ui/react-select": "latest",
-    "@radix-ui/react-separator": "latest",
-    "@radix-ui/react-slider": "latest",
-    "@radix-ui/react-slot": "latest",
-    "@radix-ui/react-switch": "latest",
-    "@radix-ui/react-tabs": "latest",
-    "@radix-ui/react-toast": "latest",
-    "@radix-ui/react-toggle": "latest",
-    "@radix-ui/react-toggle-group": "latest",
-    "@radix-ui/react-tooltip": "latest",
+    "@hookform/resolvers": "^5.2.1",
+    "@radix-ui/react-accordion": "^1.2.11",
+    "@radix-ui/react-alert-dialog": "^1.1.14",
+    "@radix-ui/react-aspect-ratio": "^1.1.7",
+    "@radix-ui/react-avatar": "^1.1.10",
+    "@radix-ui/react-checkbox": "^1.3.2",
+    "@radix-ui/react-collapsible": "^1.1.11",
+    "@radix-ui/react-context-menu": "^2.2.15",
+    "@radix-ui/react-dialog": "^1.1.14",
+    "@radix-ui/react-dropdown-menu": "^2.1.15",
+    "@radix-ui/react-hover-card": "^1.1.14",
+    "@radix-ui/react-label": "^2.1.7",
+    "@radix-ui/react-menubar": "^1.1.15",
+    "@radix-ui/react-navigation-menu": "^1.2.13",
+    "@radix-ui/react-popover": "^1.1.14",
+    "@radix-ui/react-progress": "^1.1.7",
+    "@radix-ui/react-radio-group": "^1.3.7",
+    "@radix-ui/react-scroll-area": "^1.2.9",
+    "@radix-ui/react-select": "^2.2.5",
+    "@radix-ui/react-separator": "^1.1.7",
+    "@radix-ui/react-slider": "^1.3.5",
+    "@radix-ui/react-slot": "^1.2.3",
+    "@radix-ui/react-switch": "^1.2.5",
+    "@radix-ui/react-tabs": "^1.1.12",
+    "@radix-ui/react-toast": "^1.2.14",
+    "@radix-ui/react-toggle": "^1.1.9",
+    "@radix-ui/react-toggle-group": "^1.1.10",
+    "@radix-ui/react-tooltip": "^1.2.7",
     "@supabase/ssr": "^0.6.1",
     "@supabase/supabase-js": "^2.53.0",
     "autoprefixer": "^10.4.21",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
-    "cmdk": "latest",
+    "cmdk": "^1.1.1",
     "date-fns": "4.1.0",
-    "embla-carousel-react": "latest",
-    "input-otp": "latest",
-    "lucide-react": "^0.454.0",
-    "next": "15.2.4",
-    "next-themes": "latest",
-    "react": "^19.1.0",
-    "react-day-picker": "latest",
-    "react-dom": "^19.1.0",
+    "embla-carousel-react": "^8.6.0",
+    "input-otp": "^1.4.2",
+    "lucide-react": "^0.536.0",
+    "next": "15.4.5",
+    "next-themes": "^0.4.6",
+    "react": "^19.1.1",
+    "react-day-picker": "^9.8.1",
+    "react-dom": "^19.1.1",
     "react-dropzone": "^14.3.8",
-    "react-hook-form": "latest",
+    "react-hook-form": "^7.62.0",
     "react-hot-toast": "^2.5.2",
-    "react-resizable-panels": "latest",
-    "recharts": "latest",
-    "sonner": "latest",
-    "tailwind-merge": "^2.6.0",
+    "react-resizable-panels": "^3.0.4",
+    "recharts": "^3.1.0",
+    "sonner": "^2.0.7",
+    "tailwind-merge": "^3.3.1",
     "tailwindcss-animate": "^1.0.7",
-    "vaul": "latest",
-    "zod": "^3.25.76"
+    "vaul": "^1.1.2",
+    "zod": "^4.0.14"
   },
   "devDependencies": {
-    "@types/node": "^22.16.5",
-    "@types/react": "^19.1.8",
-    "@types/react-dom": "^19.1.6",
+    "@types/node": "^24.1.0",
+    "@types/react": "^19.1.9",
+    "@types/react-dom": "^19.1.7",
     "@typescript-eslint/eslint-plugin": "^8.38.0",
     "@typescript-eslint/parser": "^8.38.0",
     "eslint": "^9.32.0",
-    "eslint-config-next": "^15.4.4",
+    "eslint-config-next": "^15.4.5",
     "eslint-plugin-import": "^2.32.0",
     "eslint-plugin-tailwindcss": "^3.18.2",
     "postcss": "^8.5.6",
-    "tailwindcss": "^3.4.17",
-    "typescript": "^5.8.3"
+    "tailwindcss": "^4.1.11",
+    "typescript": "^5.9.2"
   }
 }
