+++ apps/server/src/controllers/auth/forgot.controller.ts
