+++ apps/server/src/controllers/auth/logout.controller.ts
@@ -1,4 +1,3 @@
-// apps/api/src/controllers/v1/auth/logout.controller.ts
 /**
  * @file logout.controller.ts
  * @description This file defines the logout controller for the application.
@@ -10,7 +9,7 @@ import { Request, Response } from "express";
 
 export const logout = (req: Request, res: Response) => {
   req.session.destroy(() => {
-    res.clearCookie("trenclad-session");
+    res.clearCookie("exa-session");
     res.status(200).json({ message: "Logged out successfully" });
   });
 };
