+++ apps/server/src/controllers/auth/signup.controller.ts
