+++ apps/server/src/controllers/auth/verify.controller.ts
