+++ apps/server/src/controllers/user/delete.controller.ts
@@ -1,7 +1,7 @@
 // apps/api/src/controllers/v1/user/delete.controller.ts
 
 import { Request, Response } from "express";
-import { prisma } from "../../../config/prisma";
+import { prisma } from "../../config/prisma";
 
 export const deleteUser = async (req: Request, res: Response) => {
   const userId = req.session.userId;
