+++ apps/server/src/controllers/user/update.controller.ts
@@ -1,7 +1,7 @@
 // apps/api/src/controllers/v1/user/update.controller.ts
 
 import { Request, Response } from "express";
-import { prisma } from "../../../config/prisma";
+import { prisma } from "../../config/prisma";
 
 export const updateUser = async (req: Request, res: Response) => {
   const userId = req.session.userId;
