--- apps/server/src/routes/v2/index.ts
@@ -1,12 +1,0 @@
-import { Router } from "express";
-
-const router = Router();
-
-// Mounting routes
-
-// D
-router.get("/", (_req, res) => {
-  res.json({ message: "Welcome to API v1 🚀 (internal)" });
-});
-
-export default router;
